% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sediment.R
\name{btbr_batch_distribution}
\alias{btbr_batch_distribution}
\title{Batched Frequencies}
\usage{
btbr_batch_distribution(data, value, ...)
}
\arguments{
\item{data}{A \code{data.frame}.}

\item{value}{A \code{numeric} unquoted column.}

\item{...}{Values to pass to \link[fitdistrplus]{fitdist}}
}
\value{
A \code{data.frame} with return intervals and associated value for Weibull,
Log-pearson type III, GEV, Normal and Lognormal distributions.
}
\description{
This function fits a variable (e.g., flow, snow, rain, etc) to a univariate distribution
function, e.g. Weibull,Log-pearson type III, GEV, Normal and Lognormal distributions.
It is up to the user to decide which distribution function to use and whether any pre-processing is necessary (outliers, mixed populations, etc).
}
\examples{
granitic_dist <- btbr_batch_distribution(data \%>\% dplyr::filter(geology == 'Granitic'),
                                         value = value_tons_mi2_yr,
                                         method = 'mge')


sedimentary_dist <- btbr_batch_distribution(data \%>\% dplyr::filter(geology == 'Sedimentary'),
                                            value = value_tons_mi2_yr,
                                            method = 'mge')

}
