% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sediment.R
\name{btbr_brm_sediment}
\alias{btbr_brm_sediment}
\title{Fit Bayesian Linear Model}
\usage{
btbr_brm_sediment(btbr_rs, linear = TRUE)
}
\arguments{
\item{btbr_rs}{A previously created \link[btbr]{btbr_sediment_randomsamples} object.}

\item{linear}{Logical. Whether to run without spline or not, TRUE (default).}
}
\value{
A \code{brmsfit} model object.
}
\description{
This function fits a \link[brms]{brm} model (see details below) to previously created \link[btbr]{btbr_sediment_randomsamples}.
}
\details{
Below are the model parameter priors for the linear model. The nonlinear model using flat priors only.
\deqn{Y_i \sim \text{Beta}(\mu_i, \phi) \\
\text{logit}(\mu_i) = \beta_0 + \beta_1\text{Natural} + \beta_2\text{Roads} + \beta_3\text{Geology} \\
\beta_0=\text{Normal}(0,1) \\
\beta_1,\beta_2=\text{Normal}(0,2) \\
\beta_3=\text{Normal}(0,1) \\
\phi = \text{Gamma}(0.01, 0.01)}
}
\examples{
\dontrun{
data <- btbr_tss()

granitic_dist <- btbr_batch_distribution(data \%>\% dplyr::filter(geology == 'Granitic'),
                                         value = value_tons_mi2_yr,
                                         method = 'mge')


sedimentary_dist <- btbr_batch_distribution(data \%>\% dplyr::filter(geology == 'Sedimentary'),
                                            value = value_tons_mi2_yr,
                                            method = 'mge')

btbr_rs <- btbr_sediment_randomsamples(usfs = TRUE,
                                       sedimentary_dist = sedimentary_dist[['lpearson']],
                                       granitic_dist = granitic_dist[['lognorm']])

btbr_sedmod <- btbr_brm_sediment(btbr_rs)
}
}
