% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{btbr_pp}
\alias{btbr_pp}
\title{Get Posterior Predictive (PP)}
\usage{
btbr_pp(data, btbr_brm, indicator)
}
\arguments{
\item{btbr_brm}{A previously created \link[btbr_brm]{brbr} model.}

\item{indicator}{A character defining the type of indicator, e.g. 'sediment', 'temperature'.}

\item{A}{data.frame with appropriate covariates to predict with from previous model, e.g. \code{natural_erosion}, \code{spec_delFS}, \code{ig_or_not}, etc.}
}
\value{
A data.frame.
}
\description{
This function takes the results from the \link[btbr_brm]{brbr} model and performs posterior predictive (PP) summaries for
each Hydrological Unit Code (HUC12). The PP distribution is then split into three categories: Functioning Acceptably (FA),
Functioning at Acceptable Risk (FAR), and Functioning at Unacceptable Risk (FUR) where the cutoffs are 0-0.1 (FA), 0.1-0.3 (FAR), and
0.3-1 (FAR).
}
\note{
The \code{indicator} argument is used to split the PP into custom thresholds previously defined.
}
