% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sediment.R
\name{btbr_sediment_randomsamples}
\alias{btbr_sediment_randomsamples}
\title{Randomly Sample Total Suspended Sediment (TSS) Distributions}
\usage{
btbr_sediment_randomsamples(usfs = TRUE, sedimentary_dist, granitic_dist)
}
\arguments{
\item{sedimentary_dist}{A previously created \link[btbr]{btbr_batch_distribution} list with selected distribution.}

\item{granitic_dist}{A previously created \link[btbr]{btbr_batch_distribution} list with selected distribution.}

\item{type}{A logical. Whether to use USFS land GRAIP_Lite National Model Runs. TRUE (default)}
}
\value{
A \link[dplyr]{tibble} with new variable \code{proportion}.
}
\description{
This is a \strong{very} specific function used for the BTB model so likely won't be useful for others. The outputs of this function are used to
for proportion of road-related sediment to natural erosion in the Bayesian linear model.
}
\examples{

data <- btbr_tss()

granitic_dist <- btbr_batch_distribution(data \%>\% dplyr::filter(geology == 'Granitic'),
                                         value = value_tons_mi2_yr,
                                         method = 'mge')


sedimentary_dist <- btbr_batch_distribution(data \%>\% dplyr::filter(geology == 'Sedimentary'),
                                            value = value_tons_mi2_yr,
                                            method = 'mge')

btbr_rs <- btbr_sediment_randomsamples(usfs = TRUE,
                                       sedimentary_dist = sedimentary_dist[['lpearson']],
                                       granitic_dist = granitic_dist[['lognorm']])
}
