% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\name{btbr_temphuc_intersection}
\alias{btbr_temphuc_intersection}
\title{Temp Helper}
\usage{
btbr_temphuc_intersection(data, btb_hucs_og)
}
\arguments{
\item{data}{sf LINESTRING object. Temperature data.}

\item{btb_hucs_og}{Original FS only object.}
}
\value{
A sf object
}
\description{
Temp Helper
}
