% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sediment.R
\name{btbr_tss}
\alias{btbr_tss}
\title{Get Total Suspended Sediment (TSS) Yields}
\usage{
btbr_tss(geology = "all")
}
\arguments{
\item{geology}{A character option for returning geology type, e.g. 'all' (default), 'granitic', 'sedimentary'.}
}
\value{
A \link[dplyr]{tibble} with columns \code{value_tons_mi2_yr} and \code{value_tons_km2_yr} and associated attributes, e.g. area, total_samples, etc.
}
\description{
This function returns Total Suspended Sediment (TSS) yields in Tons/mi\if{html}{\out{<sup>}}2\if{html}{\out{</sup>}}/yr and Kg/km\if{html}{\out{<sup>}}2\if{html}{\out{</sup>}}/yr for different geology types.
These observations are then used to generate parameters for theoretical probability distribution functions (PDF) used in the
Bull Trout Baseline model.
}
\details{
The \code{geology} methods currently available:
\itemize{
\item  \strong{granitic}: These values are taken from \insertCite{kirchner2001mountain}{btbr} and 'personal communication' from \insertCite{smith1984suspended}{btbr}.
\item  \strong{sedimentary}: These values are taken from \insertCite{region1tss2024}{btbr}.
}
}
\references{
{
\insertAllCited{}
}
}
