% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_gl.R
\name{fetchGL}
\alias{fetchGL}
\title{Fetch GRAIP-Lite Data}
\usage{
fetchGL(gdb, ...)
}
\arguments{
\item{gdb}{A \code{character} vector of the GDB(s), e.g. \code{'Deschutes'}.}

\item{...}{Arguments to pass to \link{get_GL}.}
}
\value{
A list.
}
\description{
This is a higher level wrapper around the \link{get_GL} and \link{get_GL_layers}
functions. This function can fetch multiple File Geodatabases (GDB) and returns all the layers within the GDB.
}
\examples{
\dontrun{

# fetch R1 GRAIP-Lite run

GL <- fetchGL(gdb = c('R1'), quiet = TRUE)

}
}
\seealso{
\code{get_GL()} \code{get_GL_layers()}
}
\author{
Josh Erickson
}
