% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_gl.R
\name{get_GL_layers}
\alias{get_GL_layers}
\title{Get GL Layers}
\usage{
get_GL_layers(gdb, version = "v3")
}
\arguments{
\item{gdb}{A \code{character} of the GDB, e.g. \code{'R1'}.}

\item{version}{A \code{character} indicating the data version, e.g. \code{'v3'}.}
}
\value{
A list of metadata about the GDB
}
\description{
Get GL Layers
}
\note{
Refer to \code{\link{get_GL}} for information on File Geodatabase (GDB) availability.
}
\examples{
\dontrun{
GL_layers <- get_GL_layers('R1')
}

}
\author{
Josh Erickson
}
