% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemicals.R
\name{get_wcc}
\alias{get_wcc}
\title{Get WCC}
\usage{
get_wcc(
  local = TRUE,
  conda_path =
    "C:\\\\Program Files\\\\ArcGIS\\\\Pro\\\\bin\\\\Python\\\\envs\\\\arcgispro-py3"
)
}
\arguments{
\item{conda_path}{A path to an arcgispro-py3 conda env.}
}
\value{
data.frame with WCC attributes in R1.
}
\description{
Uses corporate SDE to get most recent Watershed Condition Class (WCC) information.
}
