% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{proportion_function}
\alias{proportion_function}
\title{Posterior Predictive Percentage}
\usage{
proportion_function(data, indicator)
}
\arguments{
\item{data}{A previously created posterior prediction.}

\item{indicator}{A character defining the type of indicator, e.g. 'sediment', 'temperature'.}
}
\value{
A data.frame
}
\description{
This function creates the custom splits in the PP.
}
