% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sediment.R
\name{btbr_brm_sediment}
\alias{btbr_brm_sediment}
\title{Fit Bayesian Linear Model}
\usage{
btbr_brm_sediment(btbr_rs, linear = TRUE)
}
\arguments{
\item{btbr_rs}{A previously created \link[btbr_sediment_randomsamples]{btbr} object.}

\item{linear}{Logical. Whether to run without spline or not, TRUE (default).}
}
\value{
A \code{brmsfit} model object.
}
\description{
Fit Bayesian Linear Model
}
\examples{
data <- btbr_tss()

granitic_dist <- btbr_batch_distribution(data \%>\% dplyr::filter(geology == 'Granitic'),
                                         value = value_tons_mi2_yr,
                                         method = 'mge')


sedimentary_dist <- btbr_batch_distribution(data \%>\% dplyr::filter(geology == 'Sedimentary'),
                                            value = value_tons_mi2_yr,
                                            method = 'mge')

btbr_rs <- btbr_sediment_randomsamples(usfs = TRUE,
                                       sedimentary_dist = sedimentary_dist[['lpearson']],
                                       granitic_dist = granitic_dist[['lognorm']])
btbr_sedmod <- btbr_brm(btbr_rs)
}
