% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\name{btbr_brm_temperature}
\alias{btbr_brm_temperature}
\title{Fit Bayesian Linear Model}
\usage{
btbr_brm_temperature(tempdata)
}
\arguments{
\item{tempdata}{A NorWest Stream Temperature data.frame with \code{S1_93_11} column name.}
}
\value{
A \code{brmsfit} model object.
}
\description{
Fit Bayesian Linear Model
}
\examples{

spokoot <- fishguts::get_NorWestStreams('SpoKoot') \%>\% st_as_sf()

huc_comid <-  read.csv('data/comid_huc12.csv')

spokoot_df <- dplyr::left_join(huc_comid, by = c('COMID' = 'comid')) \%>\%
  dplyr::filter(!is.na(huc12)) \%>\%
  sf::st_drop_geometry() \%>\%
  dplyr::filter(S1_93_11 > 0)

temp_model <- btbr_brm_temperature(spokoot_df)
}
