% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{btbr_hucs}
\alias{btbr_hucs}
\title{Get Bull Trout Baseline HUCs}
\usage{
btbr_hucs(usfs = TRUE)
}
\arguments{
\item{usfs}{A logical.}
}
\value{
A sf MULTIPOLYGON.
}
\description{
Get Bull Trout Baseline HUCs
}
