% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\name{btbr_norwest_temperature}
\alias{btbr_norwest_temperature}
\title{Get USDA-Forest Service Administration Boundaries}
\usage{
btbr_norwest_temperature(filter_geom, ...)
}
\arguments{
\item{filter_geom}{an object of class bbox, sfc or sfg used to filter query results based on a predicate function.}

\item{...}{Arguments to pass to \link\link{arcgislayers} package \code{arc_select} function.}
}
\value{
A sf object.
}
\description{
This layer represents modeled stream temperatures derived from the NorWeST point feature class (NorWest_TemperaturePoints).
NorWeST summer stream temperature scenarios were developed for all rivers and streams in the western U.S. from the more than 20,000 stream sites in
the NorWeST database where mean August stream temperatures were recorded. The resulting dataset includes stream lines (NorWeST_PredictedStreams)
and associated mid-points NorWest_TemperaturePoints) representing 1 kilometer intervals along the stream network.
Stream lines were derived from the 1:100,000 scale NHDPlus dataset (USEPA and USGS 2010; McKay et al. 2012, \insertCite{kirchner2001mountain}{btbr}).
Shapefile extents correspond to NorWeST processing units, which generally relate to 6 digit (3rd code) hydrologic
unit codes (HUCs) or in some instances closely correspond to state borders.
The line and point shapefiles contain identical modeled stream temperature results.
The two feature classes are meant to complement one another for use in different applications.
In addition, spatial and temporal covariates used to generate the modeled temperatures are included in the attribute tables at \url(https://www.fs.usda.gov/rm/boise/AWAE/projects/NorWeST/ModeledStreamTemperatureScenarioMaps.shtml).
The NorWeST NHDPlusV1 processing units include: Salmon, Clearwater, Spokoot, Missouri Headwaters, Snake-Bear, MidSnake, MidColumbia, Oregon Coast, South-Central Oregon, Upper Columbia-Yakima, Washington Coast, Upper Yellowstone-Bighorn, Upper Missouri-Marias, and Upper Green-North Platte.
The NorWeST NHDPlusV2 processing units include: Lahontan Basin, Northern California-Coastal Klamath, Utah, Coastal California, Central California, Colorado, New Mexico, Arizona, and Black Hills.

\verb{Copyright Text:} U.S. Forest Service; Rocky Mountain Research Station; Air, Water, and Aquatic Environments Program (AWAE). \url(https://www.fs.usda.gov/rm/boise/awae_home.shtml)
}
\references{
{
\insertAllCited{}
}
}
