% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sediment.R
\name{proportion_function}
\alias{proportion_function}
\title{Posterior Predictive Percentage}
\usage{
proportion_function(data)
}
\arguments{
\item{data}{A previously created posterior prediction.}
}
\value{
A data.frame
}
\description{
This function creates the splits in the PP.
}
